Refactor: move conflict detection code into pass
