+++ crates/solidity-v2/outputs/cargo/semantic/src/binder/mod.rs
@@ -146,10 +146,6 @@ impl Binder {
         self.scopes_by_file_id.get(file_id).copied()
     }
 
-    pub(crate) fn file_ids(&self) -> impl Iterator<Item = &String> {
-        self.scopes_by_file_id.keys()
-    }
-
     pub(crate) fn insert_scope(&mut self, scope: Scope) -> ScopeId {
         let scope_id = ScopeId(self.scopes.len());
 
@@ -336,7 +332,7 @@ impl Binder {
 
     // File scope resolution context
 
-    fn get_file_scope(&self, file_id: &str) -> &FileScope {
+    pub(crate) fn get_file_scope(&self, file_id: &str) -> &FileScope {
         self.scopes_by_file_id
             .get(file_id)
             .and_then(|scope_id| self.scopes.get(scope_id.0))
@@ -511,226 +507,6 @@ impl Binder {
                     )
             }
         }
-    }
-
-    // Looks for a previously-registered definition that conflicts with
-    // `new_definition` being declared under `symbol` in `scope_id`, returning
-    // the conflicting definition's `NodeId` (or `None` if the declaration is
-    // allowed).
-    //
-    // The search is bounded to the *local lexical chain*: block, function,
-    // modifier, parameters and yul scopes delegate to their enclosing local
-    // scope, but the walk stops at namespace boundaries (contract/file/struct/
-    // enum). This means shadowing a contract- or file-level member from an
-    // inner scope is permitted, matching Solidity. Function and event overloads
-    // are also permitted (see `Definition::overloads_with`).
-    pub(crate) fn find_conflicting_definition(
-        &self,
-        scope_id: ScopeId,
-        symbol: &str,
-        new_definition: &Definition,
-    ) -> Option<NodeId> {
-        let scope = self.get_scope_by_id(scope_id);
-        match scope {
-            Scope::Block(block_scope) => block_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition))
-                .or_else(|| {
-                    // Only continue into the parent when this block is a
-                    // "chained" continuation of the same lexical scope. A new
-                    // lexical scope (a real `{ }` block or for-init clause) may
-                    // legally shadow declarations in its enclosing scopes.
-                    if block_scope.is_lexical_scope {
-                        None
-                    } else {
-                        self.find_conflicting_definition(
-                            block_scope.parent_scope_id,
-                            symbol,
-                            new_definition,
-                        )
-                    }
-                }),
-            Scope::Function(function_scope) => function_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition))
-                .or_else(|| {
-                    // Continue into the parameters scope, but *not* the
-                    // enclosing contract/file scope (shadowing is allowed there).
-                    self.find_conflicting_definition(
-                        function_scope.parameters_scope_id,
-                        symbol,
-                        new_definition,
-                    )
-                }),
-            Scope::Modifier(modifier_scope) => modifier_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition)),
-            Scope::Parameters(parameters_scope) => parameters_scope
-                .lookup_definition(symbol)
-                .and_then(|existing| self.conflicting_definition(existing, new_definition)),
-            Scope::YulBlock(yul_block_scope) => yul_block_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition))
-                .or_else(|| {
-                    // Shadowing is not allowed in Yul, so continue lookup in the parent block
-                    self.find_conflicting_definition(
-                        yul_block_scope.parent_scope_id,
-                        symbol,
-                        new_definition,
-                    )
-                }),
-            Scope::YulFunction(yul_function_scope) => yul_function_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition)),
-            // Namespace scopes are only checked against their own definitions;
-            // the `Vec`-based ones may hold several entries for a symbol (eg.
-            // function/event overloads).
-            Scope::Contract(contract_scope) => contract_scope
-                .definitions
-                .get(symbol)
-                .and_then(|existing| self.first_conflicting_definition(existing, new_definition)),
-            Scope::File(file_scope) => file_scope
-                .definitions
-                .get(symbol)
-                .and_then(|existing| self.first_conflicting_definition(existing, new_definition)),
-            Scope::Enum(enum_scope) => enum_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition)),
-            Scope::Struct(struct_scope) => struct_scope
-                .definitions
-                .get(symbol)
-                .copied()
-                .and_then(|existing| self.conflicting_definition(existing, new_definition)),
-            Scope::Using(_) => None,
-        }
-    }
-
-    // Returns `Some(existing_id)` when an existing definition conflicts with the
-    // one being declared, or `None` when they may legally coexist (overloads).
-    fn conflicting_definition(
-        &self,
-        existing_id: NodeId,
-        new_definition: &Definition,
-    ) -> Option<NodeId> {
-        let existing = self.find_definition_by_id(existing_id)?;
-        if new_definition.overloads_with(existing) {
-            None
-        } else {
-            Some(existing_id)
-        }
-    }
-
-    // Returns the first of `existing_ids` that conflicts with `new_definition`.
-    fn first_conflicting_definition(
-        &self,
-        existing_ids: &[NodeId],
-        new_definition: &Definition,
-    ) -> Option<NodeId> {
-        existing_ids
-            .iter()
-            .find_map(|existing_id| self.conflicting_definition(*existing_id, new_definition))
-    }
-
-    // Detects clashes between each file's own declarations and the symbols
-    // brought into its scope through (unqualified) default imports.
-    //
-    // Unlike aliased or deconstructed imports — which register a local
-    // definition and are therefore caught while collecting definitions — a
-    // default import (`import "file";`) makes *all* of the imported file's
-    // symbols available without a local definition. A clash can only be
-    // detected once every file scope is populated, so this runs as a second
-    // step after all files have been visited.
-    //
-    // Returns the list of `(file_id, definition_id)` pairs identifying each
-    // local definition that conflicts with a default-imported one, ordered
-    // deterministically (by file, then symbol name).
-    pub(crate) fn find_default_import_conflicts(&self) -> Vec<(String, NodeId)> {
-        let mut conflicts: Vec<(String, NodeId)> = Vec::new();
-
-        let mut file_ids: Vec<&String> = self.file_ids().collect();
-        file_ids.sort();
-
-        for file_id in file_ids {
-            let file_scope = self.get_file_scope(file_id);
-
-            let imported_scopes = self.transitive_default_imports(file_scope);
-            if imported_scopes.is_empty() {
-                continue;
-            }
-
-            let mut symbols: Vec<&String> = file_scope.definitions.keys().collect();
-            symbols.sort();
-
-            for symbol in symbols {
-                let imported: Vec<NodeId> = imported_scopes
-                    .iter()
-                    .flat_map(|scope| scope.lookup_symbol(symbol))
-                    .collect();
-                if imported.is_empty() {
-                    continue;
-                }
-
-                for &local_id in &file_scope.definitions[symbol] {
-                    let Some(local_definition) = self.find_definition_by_id(local_id) else {
-                        continue;
-                    };
-                    let conflicts_with_import = imported.iter().any(|imported_id| {
-                        self.find_definition_by_id(*imported_id)
-                            .is_some_and(|imported| !local_definition.overloads_with(imported))
-                    });
-                    if conflicts_with_import {
-                        conflicts.push((file_id.clone(), local_id));
-                    }
-                }
-            }
-        }
-
-        conflicts
-    }
-
-    // Collects the file scopes reachable through the (transitive) default
-    // imports of `file_scope`, excluding `file_scope` itself. Mutually-recursive
-    // imports are handled by the `visited` set, which is seeded with the
-    // starting file so its own scope is never collected (even when reached
-    // again through a cycle).
-    fn transitive_default_imports<'a>(&'a self, file_scope: &'a FileScope) -> Vec<&'a FileScope> {
-        let mut found = Vec::new();
-        let mut visited: HashSet<&str> = HashSet::new();
-        visited.insert(&file_scope.file_id);
-
-        let mut queue: VecDeque<&str> = file_scope
-            .imported_files
-            .iter()
-            .map(String::as_str)
-            .collect();
-
-        while let Some(imported_file_id) = queue.pop_front() {
-            if !visited.insert(imported_file_id) {
-                continue;
-            }
-            let Some(scope_id) = self.scope_id_for_file_id(imported_file_id) else {
-                continue;
-            };
-            let Scope::File(imported_scope) = self.get_scope_by_id(scope_id) else {
-                continue;
-            };
-            found.push(imported_scope);
-            queue.extend(imported_scope.imported_files.iter().map(String::as_str));
-        }
-
-        found
     }
 
     // If the resolution points to definitions that are symbols aliases (eg.
